# PERF: Implement round on the block level