--- a/pandas/core/frame.py
+++ b/pandas/core/frame.py
@@ -9915,6 +9915,7 @@ Parrot 2  Parrot       24.0
                     yield _series_round(vals, decimals[col])
                 except KeyError:
                     yield vals
+            df._clear_item_cache()
 
         def _series_round(ser: Series, decimals: int) -> Series:
             if is_integer_dtype(ser.dtype) or is_float_dtype(ser.dtype):
@@ -9922,6 +9923,9 @@ Parrot 2  Parrot       24.0
             return ser
 
         nv.validate_round(args, kwargs)
+
+        new_cols = None
+        new_mgr = None
 
         if isinstance(decimals, (dict, Series)):
             if isinstance(decimals, Series) and not decimals.index.is_unique:
@@ -9932,14 +9936,18 @@ Parrot 2  Parrot       24.0
                 raise TypeError("Values in decimals must be integers")
             new_cols = list(_dict_round(self, decimals))
         elif is_integer(decimals):
-            # Dispatch to Series.round
-            new_cols = [_series_round(v, decimals) for _, v in self.items()]
+            # Dispatch to Block.round
+            new_mgr = self._mgr.apply("round", decimals=decimals)
         else:
             raise TypeError("decimals must be an integer, a dict-like or a Series")
 
-        if len(new_cols) > 0:
+        if new_cols is not None and len(new_cols) > 0:
             return self._constructor(
                 concat(new_cols, axis=1), index=self.index, columns=self.columns
+            ).__finalize__(self, method="round")
+        elif new_mgr is not None:
+            return self._constructor(
+                new_mgr, index=self.index, columns=self.columns
             ).__finalize__(self, method="round")
         else:
             return self.copy(deep=False)