--- a/pandas/core/internals/array_manager.py
+++ b/pandas/core/internals/array_manager.py
@@ -323,6 +323,9 @@ class BaseArrayManager(DataManager):
             cond=cond,
         )
 
+    def round(self: T, decimals: int, using_cow: bool = False) -> T:
+        return self.apply_with_block("round", decimals=decimals, using_cow=using_cow)
+
     def setitem(self: T, indexer, value) -> T:
         return self.apply_with_block("setitem", indexer=indexer, value=value)