--- a/pandas/core/internals/blocks.py
+++ b/pandas/core/internals/blocks.py
@@ -1478,10 +1478,7 @@ class Block(PandasObject):
                 # is no-op
                 # https://github.com/numpy/numpy/blob/486878b37fc7439a3b2b87747f50db9b62fea8eb/numpy/core/src/multiarray/calculation.c#L625-L636
                 values = values.copy()
-        return self.make_block_same_class(
-            values,
-            refs=refs,
-        )
+        return self.make_block_same_class(values, refs=refs)
 
     # ---------------------------------------------------------------------
     # Abstract Methods Overridden By EABackedBlock and NumpyBlock